event_management/requirements.txt
Django>=4.2.0,<5.0.0
django-debug-toolbar>=4.2.0
